+++ CHANGELOG.md
@@ -3,6 +3,25 @@ All notable changes to this project will be documented in this file.
 
 ## 4.0 Changes
 - Remove support for ThreeDSecure 1.x
+- Update WebPaymentRequestModel
+-- Add IsPayByLink
+-- Add PrimaryAccountDetails block
+-- Remove ClientIpAddress, ClientUserAgent, PartnerServiceFee
+-- Change YourPaymentMetaData to Dictionary string-object
+- Add GetWebPaymentResponseModel extending WebPaymentRequestModel that includes the attribtues returned in response to Transactions.Get that are not relevant for calls to create a web payment session
+-- CompanyName
+-- PaymentSuccessUrl
+-- PaymentCancelUrl
+-- Reference
+-- Response
+-- Status
+-- TransactionType (set on request by which interface was used)
+-- Receipt
+-- AllowedCardTypes (new)
+-- IsThreeDSecureTwo (new)
+-- NoOfAuthAttempts (new)
+- Transactions.Get and GetByReceipt now return GetWebPaymentResponseModel
+- Remove Line1/Line2/Line3 from WebPaymentCardAddress (use Address1/Address2/Address3 instead)
 
 ## 3.4 Changes on 2022-07-18
 - Allow primary account details to be set on ResumeThreeDSecureTwoModel and CompleteThreeDSecureTwoModel for MCC 6012 transactions